LPS-203854 Manage ServiceTracker from JSONWebServiceActionsManagerImpl
